+++ cartridge/lua-api/get-topology.lua
@@ -3,8 +3,10 @@
 -- @module cartridge.lua-api.get-topology
 
 local fun = require('fun')
+local yaml = require('yaml')
 local membership = require('membership')
 
+local rpc = require('cartridge.rpc')
 local utils = require('cartridge.utils')
 local roles = require('cartridge.roles')
 local failover = require('cartridge.failover')
@@ -68,6 +70,13 @@ local function get_topology()
     -- OperationError doesn't influence observing topology
     if state == 'InitError' or state == 'BootError' then
         return nil, err
+    end
+
+    if state == 'Unconfigured' and rpc.is_proxy_call_possible() then
+        local res = rpc.proxy_call('_G.__proxy_get_topology')
+        if res ~= nil then
+            return yaml.decode(res)
+        end
     end
 
     local members = membership.members()
@@ -225,6 +234,14 @@ local function get_topology()
         servers = servers,
         replicasets = replicasets,
     }
+end
+
+_G.__proxy_get_topology = function()
+    local res, err = get_topology()
+    if not res then
+        return nil, err
+    end
+    return yaml.encode(res)
 end
 
 return {